shuffle all the things
