+++ src/CribblyBackend.Core/Extensions/ServiceCollectionExtensions.cs
@@ -1,4 +1,5 @@
-using System;
+using CribblyBackend.Core.Divisions;
+using CribblyBackend.Core.Divisions.Services;
 using CribblyBackend.Core.Games.Services;
 using CribblyBackend.Core.Players.Services;
 using CribblyBackend.Core.Teams.Services;
