+++ src/CribblyBackend.DataAccess/Divisions/DivisionRepository.cs
@@ -1,19 +1,11 @@
-using System.Collections.Generic;
 using System.Data;
-using System.Linq;
 using System.Threading.Tasks;
-using CribblyBackend.DataAccess.Models;
+using CribblyBackend.Core.Divisions.Models;
+using CribblyBackend.Core.Divisions.Repositories;
 using Dapper;
 
-namespace CribblyBackend.DataAccess.Repositories
+namespace CribblyBackend.DataAccess.Divisions
 {
-    public interface IDivisionRepository
-    {
-        void GetById(int id);
-        void Update(Division diivision);
-        Task<Division> Create(Division division);
-        void Delete(Division division);
-    }
     public class DivisionRepository : IDivisionRepository
     {
         private readonly IDbConnection _connection;
@@ -29,8 +21,8 @@ namespace CribblyBackend.DataAccess.Repositories
         public async Task<Division> Create(Division division)
         {
             var result = await _connection.ExecuteAsync(
-                @"INSERT INTO Divisions(Name) VALUES (@Name)", 
-                new {Name = division.Name}
+                @"INSERT INTO Divisions(Name) VALUES (@Name)",
+                new { Name = division.Name }
             );
 
             return division;
