+++ src/CribblyBackend.DataAccess/Extensions/ServiceCollectionExtensions.cs
@@ -1,10 +1,11 @@
+using CribblyBackend.Core.Divisions.Repositories;
 using CribblyBackend.Core.Games.Repositories;
 using CribblyBackend.Core.Players.Repositories;
 using CribblyBackend.Core.Teams.Repositories;
 using CribblyBackend.Core.Tournaments.Repositories;
+using CribblyBackend.DataAccess.Divisions;
 using CribblyBackend.DataAccess.Games.Repositories;
 using CribblyBackend.DataAccess.Players.Repositories;
-using CribblyBackend.DataAccess.Repositories;
 using CribblyBackend.DataAccess.Teams.Repositories;
 using CribblyBackend.DataAccess.Tournaments.Repositories;
 using Microsoft.Extensions.DependencyInjection;
